Upload layer 1 to tilemap
This commit uploads the foreground of the square level to VRAM at $0800 and sets it as the layer 1 tilemap. Layer 1 isn't visible and layer 2 is completely ignored, but I spent over an hour trying to figure out why my DMAs were just not doing anything. Also, for some weird reason, NSLVL is offset 2 bytes, so a dirty hack is included to ignore that. Layer 2 is coming next commit, then I might actually display the tilemaps on the screen.